refactor(Core): Split up rust code into different modules
